@@ -24,18 +24,18 @@
   "homepage": "http://ummels.github.io/jekyll-aws-comments",
   "private": true,
   "dependencies": {
-    "js-md5": "^0.4.0",
-    "moment": "^2.11.2",
-    "octokat": "^0.6.4"
+    "js-md5": "^0.7.3",
+    "moment": "^2.24.0",
+    "octokat": "^0.10.0"
   },
   "devDependencies": {
     "del": "^2.2.0",
-    "gulp": "^3.9.1",
-    "gulp-awslambda": "^0.4.0",
+    "gulp": "^4.0.2",
+    "gulp-awslambda": "^0.5.0",
     "gulp-install": "^1.1.0",
-    "gulp-zip": "^4.0.0"
+    "gulp-zip": "^5.0.1"
   },
   "engines": {
-    "node": ">= 6.10"
+    "node": ">= 12.1"
   }
 }
